+++ integration_tests/conftest.py
@@ -4,7 +4,10 @@ import asyncio
 import io
 import multiprocessing
 import os
+import signal
+import subprocess
 import time
+from pathlib import Path
 from urllib.error import URLError
 from urllib.request import urlopen
 
@@ -12,6 +15,8 @@ import boto3
 import pytest
 import uvicorn
 from botocore.exceptions import ClientError
+
+from .celery_example import MultilineJsonFileGateway
 
 
 def pytest_sessionstart(session):
@@ -100,6 +105,37 @@ async def fastapi_example_app():
         yield f"http://localhost:{port}"
     finally:
         p.terminate()
+
+
+@pytest.fixture(scope="session")
+def celery_worker(tmp_path_factory):
+    log_file = str(tmp_path_factory.mktemp("pytest-celery") / "celery.log")
+    p = subprocess.Popen(
+        [
+            "celery",
+            "-A",
+            "integration_tests.celery_example",
+            "worker",
+            "-c",
+            "1",
+            # "-P",  enable when using the debugger
+            # "solo"
+        ],
+        start_new_session=True,
+        stdout=subprocess.PIPE,
+        # optionally add "CLEAN_PYTHON_TEST_DEBUG": "5679" to enable debugging
+        env={"CLEAN_PYTHON_TEST_LOGGING": log_file, **os.environ},
+    )
+    try:
+        yield MultilineJsonFileGateway(Path(log_file))
+    finally:
+        p.send_signal(signal.SIGQUIT)
+
+
+@pytest.fixture
+def celery_task_logs(celery_worker):
+    celery_worker.clear()
+    return celery_worker
 
 
 @pytest.fixture(scope="session")